move (fake) assessment to server
