+++ frontend/src/components/DomainCard.tsx
@@ -92,7 +92,8 @@ export const ImpossibleBanner = ({ isPossible, isAvailable, isCheap }: { isPossi
 
 export const DomainCard = (assessment: DomainAssessment) => {
     const [isLiked, setIsLiked] = useState(false);
-    const { domain, isPossible, isAvailable, isCheap, evoc, brev, grep, goog, pron, spel, verb } = assessment;
+    const { domain, isPossible, isAvailable, isCheap, scores } = assessment;
+    const { evoc, brev, grep, goog, pron, spel, verb } = scores;
     const totalScore = evoc + brev + grep + goog + pron + spel + verb;
 
     const isValid = isPossible && isAvailable && isCheap
