+++ frontend/src/serverCalls.ts
@@ -1,3 +1,5 @@
+import { DomainAssessment } from "shared/types";
+
 const SERVER_PATH = import.meta.env.VITE_SERVER_URL;
 
 type UserInput = {
@@ -10,7 +12,7 @@ type UserInput = {
 
 export const sendInputsAndReturnDomains = async (
   userInput: UserInput
-): Promise<string[]> => {
+): Promise<DomainAssessment[]> => {
   const response = await fetch(`${SERVER_PATH}/find-domains`, {
     method: "POST",
     body: JSON.stringify({ userInput }),
@@ -19,13 +21,14 @@ export const sendInputsAndReturnDomains = async (
     },
   });
   const json = await response.json();
-  const domainList: string[] = json.domains;
-  console.log("The server response was:", domainList);
-  return domainList;
+  const results: DomainAssessment[] = json.domainAssessments;
+  console.log("The json response was:", json);
+  console.log("The processed response was:", results);
+  return results;
 };
 
 export const checkHeartbeat = async (): Promise<void> => {
   const response = await fetch(`${SERVER_PATH}/heartbeat`);
   const json = await response.json();
-  console.log("The server response was:", json);
+  console.log("Warmup ping sent. Server response:", json.message);
 };
