+++ frontend/tsconfig.json
@@ -32,5 +32,5 @@
     },
     "rootDir": ".."
   },
-  "include": ["src/**/*", "vite.config.ts", "../shared/**/*", "graveyard/*"]
+  "include": ["src/**/*", "vite.config.ts", "../shared/**/*"]
 }
